Move some utility functions around
