+++ reflection.go
@@ -1,0 +1,18 @@
+package mars
+
+import (
+	"reflect"
+)
+
+// Return the reflect.Method, given a Receiver type and Func value.
+func findMethod(recvType reflect.Type, funcVal reflect.Value) *reflect.Method {
+	// It is not possible to get the name of the method from the Func.
+	// Instead, compare it to each method of the Controller.
+	for i := 0; i < recvType.NumMethod(); i++ {
+		method := recvType.Method(i)
+		if method.Func.Pointer() == funcVal.Pointer() {
+			return &method
+		}
+	}
+	return nil
+}
